create ABC for Extractors
